addressing Ryan's comments
