+++ inc/packages/admin/namespace.php
@@ -70,8 +70,8 @@ function handle_did_during_ajax( $result, $action, $args ) {
 	}
 
 	wp_cache_set( ACTION_INSTALL_DID, $did );
-	Updater\add_package_to_release_cache( $did );
-	add_filter( 'http_request_args', 'FAIR\\Updater\\maybe_add_accept_header', 20, 2 );
+	Packages\add_package_to_release_cache( $did );
+	add_filter( 'http_request_args', 'FAIR\\Packages\\maybe_add_accept_header', 20, 2 );
 
 	return (object) Packages\get_update_data( $did );
 }
