+++ inc/packages/namespace.php
@@ -16,6 +16,7 @@ use WP_Upgrader;
 const SERVICE_ID = 'FairPackageManagementRepo';
 const CONTENT_TYPE = 'application/json+fair';
 const CACHE_LIFETIME = 12 * HOUR_IN_SECONDS;
+const RELEASE_PACKAGES_CACHE_KEY = 'fair-release-packages';
 
 // phpcs:disable WordPress.NamingConventions.ValidVariableName
 
@@ -624,4 +625,51 @@ function rename_source_selection( string $source, string $remote_source, WP_Upgr
 
 	return trailingslashit( $new_source );
 }
+
+/**
+ * Add FAIR ReleaseDocument data to cache.
+ *
+ * @param string $did DID.
+ * @return void
+ */
+function add_package_to_release_cache( string $did ) : void {
+	if ( empty( $did ) ) {
+		return;
+	}
+	$releases = wp_cache_get( RELEASE_PACKAGES_CACHE_KEY ) ?: [];
+	$releases[ $did ] = get_latest_release_from_did( $did );
+	wp_cache_set( RELEASE_PACKAGES_CACHE_KEY, $releases );
+}
+
+/**
+ * Maybe add accept header for release asset package binary.
+ *
+ * ReleaseDocument artifact package content-type will be application/octet-stream.
+ * Only for GitHub release assets.
+ *
+ * @param array  $args Array of http args.
+ * @param string $url  Download URL.
+ *
+ * @return array
+ */
+function maybe_add_accept_header( $args, $url ) : array {
+	$releases = wp_cache_get( RELEASE_PACKAGES_CACHE_KEY ) ?: [];
+
+	if ( ! str_contains( $url, 'api.github.com' ) ) {
+		return $args;
+	}
+
+	foreach ( $releases as $release ) {
+		if ( $url === $release->artifacts->package[0]->url ) {
+			$content_type = $release->artifacts->package[0]->{'content-type'};
+			if ( $content_type === 'application/octet-stream' ) {
+				$args = array_merge( $args, [ 'headers' => [ 'Accept' => $content_type ] ] );
+				break;
+			}
+		}
+	}
+
+	return $args;
+}
+
 // phpcs:enable
